worked on the controller
